+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+### Bugs Fixed
+
+* [#216](https://github.com/clojure-emacs/orchard/issues/216): Fix some pagination issues with inspector navigation.
+  * `orchard.inspect/down` fn is now able to go beyond the current page. `next-sibling` and `previous-sibling` fns work in all pages.
+
 ## 0.23.0 (2024-03-03)
 
 ## Changes